A rogue branch for iferret light.  Seems to work on 64 bit

+++ iferret_open_fd.c
@@ -29,7 +29,7 @@ void iferret_open_fd_add(iferret_t *iferret, int pid, int fd, char *filename,
   else {
     // no table yet for that pid.  create it first.  
     ofd_table = int_int_hashtable_new();
-    int_int_hashtable_add(iferret->open_fd_table, pid, (uint32_t) ofd_table);
+    int_int_hashtable_add(iferret->open_fd_table, pid, (uint64_t) ofd_table);
   }
   if (int_int_hashtable_mem(ofd_table, fd)) {
     // Hey, we already have an entry for this pid/fd. wtf? shouldnt happen
@@ -44,7 +44,7 @@ void iferret_open_fd_add(iferret_t *iferret, int pid, int fd, char *filename,
     exit(1);
   }
   // finally add ptr to open fd data to pid/fd table 
-  int_int_hashtable_add(ofd_table, fd, (uint32_t) ofd_info);
+  int_int_hashtable_add(ofd_table, fd, (uint64_t) ofd_info);
 }
